backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.28.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1
python-dotenv==1.0.0
httpx==0.24.1
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7
websockets==12.0
aiohttp==3.8.6
faker==19.13.0
APScheduler==3.10.4
sendgrid==6.10.0
supabase==2.2.0
pynacl==1.5.0
firebase-admin==6.2.0
python-dateutil==2.8.2
geopy==2.4.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-socketio==5.8.0